testReduceFromPartialMem: Add Shuffle IO error assertion to test
