+++ hadoop-mapreduce-project/hadoop-mapreduce-client/hadoop-mapreduce-client-shuffle/src/test/java/org/apache/hadoop/mapred/TestShuffleHandler.java
@@ -17,16 +17,15 @@
  */
 package org.apache.hadoop.mapred;
 
+import io.netty.channel.DefaultFileRegion;
+import org.apache.commons.compress.changes.ChangeSetPerformer;
 import org.apache.hadoop.thirdparty.com.google.common.collect.Maps;
-import io.netty.buffer.ByteBuf;
 import io.netty.channel.AbstractChannel;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelPipeline;
-import io.netty.channel.ChannelPromise;
 import io.netty.channel.socket.SocketChannel;
-import io.netty.handler.codec.http.HttpHeaders;
 import io.netty.handler.codec.http.HttpMethod;
 import io.netty.handler.codec.http.HttpRequest;
 import io.netty.handler.codec.http.HttpResponse;
@@ -318,76 +317,6 @@ public class TestShuffleHandler {
         }
         super.channelIdle(ctx, e);
       }
-    }
-  }
-
-  static class LoggingHttpResponseEncoder extends HttpResponseEncoder {
-    private final boolean logStacktraceOfEncodingMethods;
-
-    public LoggingHttpResponseEncoder(boolean logStacktraceOfEncodingMethods) {
-      this.logStacktraceOfEncodingMethods = logStacktraceOfEncodingMethods;
-    }
-
-    @Override
-    public boolean acceptOutboundMessage(Object msg) throws Exception {
-      printExecutingMethod();
-      return super.acceptOutboundMessage(msg);
-    }
-
-    @Override
-    protected void encodeInitialLine(ByteBuf buf, HttpResponse response) throws Exception {
-      LOG.debug("Executing method: {}, response: {}",
-          getExecutingMethodName(), response);
-      logStacktraceIfRequired();
-      super.encodeInitialLine(buf, response);
-    }
-
-    @Override
-    protected void encode(ChannelHandlerContext ctx, Object msg,
-        List<Object> out) throws Exception {
-      printExecutingMethod();
-      logStacktraceIfRequired();
-      super.encode(ctx, msg, out);
-    }
-
-    @Override
-    protected void encodeHeaders(HttpHeaders headers, ByteBuf buf) {
-      printExecutingMethod();
-      super.encodeHeaders(headers, buf);
-    }
-
-    @Override
-    public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise
-        promise) throws Exception {
-      printExecutingMethod();
-      super.write(ctx, msg, promise);
-    }
-
-    private void logStacktraceIfRequired() {
-      if (logStacktraceOfEncodingMethods) {
-        LOG.debug("Stacktrace: ", new Throwable());
-      }
-    }
-
-    private void printExecutingMethod() {
-      String methodName = getExecutingMethodName();
-      LOG.debug("Executing method: {}", methodName);
-    }
-
-    private String getExecutingMethodName() {
-      StackTraceElement[] stackTrace = Thread.currentThread()
-          .getStackTrace();
-      // Array items (indices):
-      // 0: java.lang.Thread.getStackTrace(...)
-      // 1: TestShuffleHandler$LoggingHttpResponseEncoder.getExecutingMethodName(...)
-      String methodName = stackTrace[2].getMethodName();
-      //If this method was called from printExecutingMethod, 
-      // we have yet another stack frame
-      if (methodName.endsWith("printExecutingMethod")) {
-        methodName = stackTrace[3].getMethodName();
-      }
-      String className = this.getClass().getSimpleName();
-      return className + "#" + methodName;
     }
   }
